reverse int64

+++ internal/domain/models/log_entry.go
@@ -7,5 +7,5 @@ type LogEntry struct {
 	Level     string
 	Message   string
 	Metadata  map[string]*structpb.Value
-	Timestamp int32
+	Timestamp int64
 }
